Moved login components into atomic pattern folders
