+++ source/components/molecules/MoreInfoExpanded.js
